revisions_1

+++ src/Karim/Developer.java
@@ -23,7 +23,7 @@ public class Developer
 	}
 
 	
-	public int getDeveloperNextAvailableHour() {
+	public double getDeveloperNextAvailableHour() {
 		return developerNextAvailableHour;
 	}
